Fix endpoint logging
The endpoint variable was used in the routes functions, but were
redefined for each endpoint.
Thus, for any endpoint, the `endpoint` variable always had the
same value: the last defined one (i.e. redirectEndpoints.write for
operator-node, and jsonProxyEndpoints.createSeed for client-node).

To fix this, we remove the `endpoint` variable and directly use
its value.