Do not call the operator for deleting prefs when there are none
If the user is already not participating, there is no need to call the
operator as it would functionnaly be a no-op.

This would be the case if it's the first time the user visits the
website and they click on the refuse all button, without having chosen
any preferences.