Add delete-ids-prefs via redirect
The deleteIdsAndPreferences function now supports both 3PC and non-3PC
variants.

Note that the response of GetIdsPrefs/PostIdsPrefs/DeleteIdsPrefs follow
the same type, which is how we're currently reading the ids/prefs
after a redirect.

Identifiers being empty + undefined preferences are used to specify
that the ids must be deleted.