Implemented the V2 Audit Log user interface.
Included the OneKey icon under the advert once the advert has loaded and the audit log data provided.

Removed the GB language file for audit to focus on getting the US one correct before copying.

Added helpers to create mock audit logs.

Included tests for the audit viewer model.

Modified the UI to accept mock audit logs for display in the UI when the real audit data is not available or as part of the demo.

Added a README.md to the audit viewer module and index.html to show how the audit viewer can be loaded in a simple HTML page.